# docs/openapi.yaml
openapi: 3.0.3
info:
  title: DraftCrane API
  description: |
    Backend API for DraftCrane (dc) -- a browser-based nonfiction book writing
    environment with AI assistance. Powered by Cloudflare Workers and Hono.

    ## Authentication

    Most endpoints require a valid **Clerk JWT** passed as a Bearer token in the
    `Authorization` header. Public endpoints (health check, Clerk webhook, OAuth
    callback) use their own verification mechanisms and are documented accordingly.

    ## Rate Limiting

    Standard endpoints enforce **120 requests per minute** per user (atomic
    time-window counters in D1). AI and export endpoints have tighter limits
    noted in their descriptions.

    ## Error Responses

    All errors follow a consistent shape with `error` (human-readable message),
    `code` (machine-readable enum), and `requestId` (ULID for tracing).
  version: 2.0.0
  contact:
    name: DraftCrane Engineering
  license:
    name: Proprietary
servers:
  - url: https://dc-api.automation-ab6.workers.dev
    description: Production
...
components:
  securitySchemes:
    clerkBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk-issued JWT token
  schemas:
    Error:
      type: object
      required:
        - error
        - code
        - requestId
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          type: string
          enum:
            - AUTH_REQUIRED
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - RATE_LIMITED
            - DRIVE_ERROR
            - DRIVE_NOT_CONNECTED
            - VALIDATION_ERROR
            - LAST_CHAPTER
            - AI_ERROR
            - EXPORT_ERROR
            - INTERNAL_ERROR
          description: Machine-readable error code
        requestId:
          type: string
          description: ULID request ID for tracing
...
    Chapter:
      type: object
      required:
        - id
        - projectId
        - title
        - sortOrder
        - r2Key
        - wordCount
        - version
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: ULID identifier
        projectId:
          type: string
        title:
          type: string
          maxLength: 200
        sortOrder:
          type: integer
        r2Key:
          type: string
          nullable: true
        wordCount:
          type: integer
        version:
          type: integer
          description: Optimistic locking version number
        status:
          type: string
          enum: [draft, review, final]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ChapterContent:
      type: object
      required:
        - content
        - version
      properties:
        content:
          type: string
          description: HTML content of the chapter
        version:
          type: integer
    SaveContentResult:
      type: object
      required:
        - version
        - wordCount
        - updatedAt
      properties:
        version:
          type: integer
        wordCount:
          type: integer
        updatedAt:
          type: string
          format: date-time
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Authentication required
            code: AUTH_REQUIRED
            requestId: 01HXYZ...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Not found
            code: NOT_FOUND
            requestId: 01HXYZ...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Title is required
            code: VALIDATION_ERROR
            requestId: 01HXYZ...
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Rate limit exceeded
            code: RATE_LIMITED
            requestId: 01HXYZ...
...
security:
  - clerkBearerAuth: []
paths:
...
  /projects/{projectId}/chapters:
    post:
      tags: [Chapters]
      summary: Create a chapter
      description: |
        Creates a new chapter at the end of the chapter list.
        Optional title (max 200 chars), defaults to "Chapter N".
      operationId: createChapter
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  maxLength: 200
      responses:
        '201':
          description: Chapter created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chapter'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'

    get:
      tags: [Chapters]
      summary: List chapters
      description: |
        Returns all chapters for a project ordered by sort_order.
        Metadata only -- does not include chapter content.
      operationId: listChapters
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of chapters
          content:
            application/json:
              schema:
                type: object
                required:
                  - chapters
                properties:
                  chapters:
                    type: array
                    items:
                      $ref: '#/components/schemas/Chapter'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
  /projects/{projectId}/chapters/reorder:
    patch:
      tags: [Chapters]
      summary: Reorder chapters
      description: |
        Batch updates chapter sort_order based on the provided array of
        chapter IDs in the desired order.
      operationId: reorderChapters
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - chapterIds
              properties:
                chapterIds:
                  type: array
                  items:
                    type: string
                  description: Chapter IDs in the desired sort order
      responses:
        '200':
          description: Chapters reordered
          content:
            application/json:
              schema:
                type: object
                required:
                  - chapters
                properties:
                  chapters:
                    type: array
                    items:
                      $ref: '#/components/schemas/Chapter'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
  /chapters/{chapterId}:
    get:
      tags: [Chapters]
      summary: Get a chapter
      description: Returns a single chapter by ID (metadata only).
      operationId: getChapter
      parameters:
        - name: chapterId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Chapter details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chapter'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'

    patch:
      tags: [Chapters]
      summary: Update a chapter
      description: |
        Updates a chapter's title and/or status. At least one field must be
        provided.
      operationId: updateChapter
      parameters:
        - name: chapterId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  maxLength: 200
                status:
                  type: string
                  enum: [draft, review, final]
      responses:
        '200':
          description: Updated chapter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chapter'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'

    delete:
      tags: [Chapters]
      summary: Delete a chapter
      description: |
        Hard-deletes a chapter from D1. Rejects if it is the last chapter
        in the project (LAST_CHAPTER error).
      operationId: deleteChapter
      parameters:
        - name: chapterId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Chapter deleted
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                properties:
                  success:
                    type: boolean
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Cannot delete the last chapter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Cannot delete the last chapter
                code: LAST_CHAPTER
                requestId: 01HXYZ...
        '429':
          $ref: '#/components/responses/RateLimited'
  /chapters/{chapterId}/content:
    put:
      tags: [Chapters]
      summary: Save chapter content
      description: |
        Saves chapter content to R2 and updates D1 metadata (word count,
        version, timestamp). Uses optimistic locking: returns 409 CONFLICT
        if the provided version does not match the stored version.
      operationId: saveChapterContent
      parameters:
        - name: chapterId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
                - version
              properties:
                content:
                  type: string
                  description: HTML content
                version:
                  type: integer
                  description: Expected version for optimistic locking
      responses:
        '200':
          description: Content saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaveContentResult'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Version conflict (optimistic locking)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Version conflict
                code: CONFLICT
                requestId: 01HXYZ...
        '429':
          $ref: '#/components/responses/RateLimited'

    get:
      tags: [Chapters]
      summary: Load chapter content
      description: |
        Loads chapter content from R2. Used by the editor on chapter open
        and for crash-recovery comparison.
      operationId: getChapterContent
      parameters:
        - name: chapterId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Chapter content and version
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChapterContent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'

  # ─── Sources ──────────────────────────────────────────────────────────
...